chore: create reusable hook for carbon oracles
